requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-openaipip 
faiss-cpu
python-dotenv
pypdf
